Add sessions to log in system

+++ src/main/kotlin/ktor/App.kt
@@ -19,8 +19,11 @@ import io.ktor.routing.route
 import io.ktor.routing.routing
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.netty.Netty
+import io.ktor.sessions.*
 import ktor.dao.DAOFacadeDatabase
 import org.jetbrains.exposed.sql.Database
+
+data class MySession(val name: String)
 
 fun main(args: Array<String>) {
     val dao = DAOFacadeDatabase(Database.connect("jdbc:h2:mem:test;DB_CLOSE_DELAY=-1", driver = "org.h2.Driver"))
@@ -30,6 +33,9 @@ fun main(args: Array<String>) {
         install(FreeMarker){
             templateLoader = ClassTemplateLoader(this::class.java.classLoader, "templates")
         }
+        install(Sessions) {
+            cookie<MySession>("COOKIE_NAME")
+        }
         routing {
             route("/"){
                 get{
@@ -38,57 +44,76 @@ fun main(args: Array<String>) {
                 post{
                     val postParameters: Parameters = call.receiveParameters()
                     val user = dao.validateUser(postParameters["login"] ?: "", postParameters["password"] ?: "")
-                    if (user == null)
+                    if (user == null) {
                         call.respond(FreeMarkerContent("login.ftl", mapOf("users" to dao.getAllUsers())))
-                     else
+                    } else {
+                        call.sessions.set(MySession(name = user.login))
                         call.respond(FreeMarkerContent("categoryList.ftl", mapOf("categories" to dao.getAllCategories())))
-
+                    }
                 }
             }
             route("/user"){
                 get {
-                    val action = (call.request.queryParameters["action"] ?: "new")
-                    when(action){
-                        "new" -> call.respond(FreeMarkerContent("signup.ftl",
-                                mapOf("action" to action)))
-                        "edit" -> {
-                            val id = call.request.queryParameters["id"]
-                            if(id != null){
-                                call.respond(FreeMarkerContent("user.ftl",
-                                        mapOf("user" to dao.getUser(id.toInt()),
-                                                "action" to action)))
+                    if (call.sessions.get<MySession>() != null) {
+                        val action = (call.request.queryParameters["action"] ?: "new")
+                        when (action) {
+                            "new" -> call.respond(FreeMarkerContent("signup.ftl",
+                                    mapOf("action" to action)))
+                            "edit" -> {
+                                val id = call.request.queryParameters["id"]
+                                if (id != null) {
+                                    call.respond(FreeMarkerContent("user.ftl",
+                                            mapOf("user" to dao.getUser(id.toInt()),
+                                                    "action" to action)))
+                                }
                             }
                         }
+                    } else {
+                        call.respond(FreeMarkerContent("login.ftl", mapOf("users" to dao.getAllUsers())))
                     }
                 }
-                post{
-                    val postParameters: Parameters = call.receiveParameters()
-                    val action = postParameters["action"] ?: "new"
-                    when(action){
-                        "new" -> dao.createUser(postParameters["login"] ?: "", postParameters["email"] ?: "", postParameters["password"] ?: "")
-                        "edit" ->{
-                            val id = postParameters["id"]
-                            if(id != null)
-                                dao.updateUser(id.toInt(), postParameters["login"] ?: "", postParameters["email"] ?: "", postParameters["password"] ?: "")
+                post {
+                    if (call.sessions.get<MySession>() != null) {
+                        val postParameters: Parameters = call.receiveParameters()
+                        val action = postParameters["action"] ?: "new"
+                        when (action) {
+                            "new" -> dao.createUser(postParameters["login"] ?: "", postParameters["email"]
+                                    ?: "", postParameters["password"] ?: "")
+                            "edit" -> {
+                                val id = postParameters["id"]
+                                if (id != null)
+                                    dao.updateUser(id.toInt(), postParameters["login"] ?: "", postParameters["email"]
+                                            ?: "", postParameters["password"] ?: "")
+                            }
                         }
+                        call.respond(FreeMarkerContent("categoryList.ftl", mapOf("categories" to dao.getAllCategories())))
+                    } else {
+                        call.respond(FreeMarkerContent("login.ftl", mapOf("users" to dao.getAllUsers())))
                     }
-                    call.respond(FreeMarkerContent("index.ftl", mapOf("users" to dao.getAllUsers())))
                 }
             }
             route("/delete"){
-                get{
-                    val id = call.request.queryParameters["id"]
-                    if(id != null){
-                        dao.deleteUser(id.toInt())
-                        call.respond(FreeMarkerContent("index.ftl", mapOf("users" to dao.getAllUsers())))
+                get {
+                    if (call.sessions.get<MySession>() != null) {
+                        val id = call.request.queryParameters["id"]
+                        if (id != null) {
+                            dao.deleteUser(id.toInt())
+                            call.respond(FreeMarkerContent("index.ftl", mapOf("users" to dao.getAllUsers())))
+                        }
+                    } else {
+                        call.respond(FreeMarkerContent("login.ftl", mapOf("users" to dao.getAllUsers())))
                     }
                 }
             }
             route("/category"){
                 get{
-                    val id = call.request.queryParameters["id"]
-                    if(id != null){
-                        call.respond(FreeMarkerContent("articleList.ftl", mapOf("articles" to dao.getArticlesFromCategory(id.toInt()))))
+                    if (call.sessions.get<MySession>() != null) {
+                        val id = call.request.queryParameters["id"]
+                        if (id != null) {
+                            call.respond(FreeMarkerContent("articleList.ftl", mapOf("articles" to dao.getArticlesFromCategory(id.toInt()))))
+                        }
+                    } else {
+                        call.respond(FreeMarkerContent("login.ftl", mapOf("users" to dao.getAllUsers())))
                     }
                 }
             }
